Remove fifo_bulk_schedule
